+++ src/actions/index.js
@@ -31,7 +31,7 @@ export const SET_LAYOUTS = 'SET_LAYOUTS';
 export const SET_BOXES_LAYOUTS = 'SET_BOXES_LAYOUTS';
 export const SET_MENU = 'SET_MENU';
 export const SET_TOPOLOGY_OPTION = 'SET_TOPOLOGY_OPTION';
-
+export const SET_ALERT = 'SET_ALERT';
 
 export function setBoxes(boxes) {
     return {
@@ -289,6 +289,13 @@ export function setTopologyOption(topologyOption) {
     return {
         type: SET_TOPOLOGY_OPTION,
         topologyOption: topologyOption
+    };
+}
+
+export function setAlert(alert) {
+    return {
+        type: SET_TOPOLOGY_OPTION,
+        alert: alert
     };
 }
