+++ src/components/Paper/PaperControl/AlertList.js
@@ -10,7 +10,7 @@ class AlertList extends Component {
     render() {
 
         return (
-            <div className={"popup-div alert-list " + (this.props.show ? '' : 'hidden')}>
+            <div className={"popup-div alert-list-wrapper " + (this.props.show ? '' : 'hidden')}>
                 <div className="top-control">
                     <div className="alerts">ALERTS</div>
                     <button onClick={this.props.clearAllAlert}>CLEAR ALL</button>
